requirements.txt
python-dotenv
pydantic
nemoguardrails
pinecone-client
datasets
openai
chainlit